requirements.txt
Werkzeug==2.0.1
flask-cors==3.0.10
openai==0.27.8
SpeechRecognition==3.8.1
numpy==1.23.5
pandas==1.5.3
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1